Update css framework build process for npm
Previously if a developer published the css framework then the
static version would not be included.

This change builds the static css file before publication, puts it
in a folder named dist (the norm for distribution) and
ensures it is included in the build. Furthermore it adds props
to the package file to instruct it which files to includes to fix
an issue in which publication was falling back to the gitignore file.

Finally the links to the git repo have been updated so when
published to npm those links will send the user to the correct
github repo and page.