Correct minor issues in react component package
The package for react components has been updated to point the
user to the correct url when published to npm.com

Dependencies have been updated to be in the correct order and
include the autosuggest package as a regular dependency
so that users of the package don't need to manually install it.

Added step to build the package before publication.