Update monorepo lint command
Updates the lint and lint-staged commands at the
top level of the project to call All sub-projects that
can be linted to ensure they are.

Uses concurrently to allow processes to run in parallel
and work with windows.

Use yarn --cwd to set the current folder instead of having
to cd to it.