Add support for links to Vue Dropdown
The previous version of dropdown only allowed the component to act like a field
and emitted a value in an 'input' event. This updated version introduces the ability
to pass in a list of options with a url for each one, so the dropdown is used as a
navigation aid.