Replace node-sass with dart sass for packages using storybook
Trying to remove all trace of node-sass to remove the need for a binary and address a security alert on node-gyp/tar. This commit changes the packages that use storybook. The webpack config passes an option to the sass loader to enforce the use of dart sass over node sass.