Deprecate React Component JS library and replace with TS
Move the old React component library out of the package structure so the TS version can replace it. Keeping it out of the package structure allows us to access the old source to migrate it as needed before deleting it. Updated all the build packages and scripts to use the new package name