Focus on searchbar when it comes into view and hide it when searching
When the searchbar came into view the user had to click on it to use it, which felt odd. This changes the behaviour to autofocus on the searchbar input when it is shown.

When a user submits a search from the masthead the searchbar is automatically hidden again to indicate a search has taken place.