Fix build issues associated with introducing icons
If an icon from the icon library is included in a react component or the docs site the build time grows many times and the output bundle for react components is almost a megabyte bigger. By putting the icons into the projects that need them the size is brought down until we can tree shake unused icons or tell the build tool not to bundle them in the output.