Add support for errors to native form components
withFormik displayed errors outside of a form component, but that means it stripped away the ability to pass an error message into a component and have it displayed. Native form components need to be able to display errors so users can use them and the upcoming redux form components can also take advantage of it without duplicating code.