Add Formik wrapper in docs around Text Inputs to fix state bug
Previously the Text Input documentation was demonstrating how to use the TextInput in a Formik form, but wasn't actually using Formik. This caused an error in that the example had no stateful container so didn't work as it should. Adding Formik to the project and creating a Formik form makes the example act as described.