Change button group behaviour
Changes the ButtonGroup component to generate markup for a collection of buttons, grouped together. All the properties of the buttons are set at the top level, and filter down and buttons are now only secondary buttons with no colour variation. The ability to include disabled buttons has also been added.