fix(Types): Rename and export ComponentWithClass
Previously the ComponentWithClass interface was not exported. This meant that it wasn't included in the distributed types (in `packages/react-component-library/dist/types`) and users of the component library could not set the `className` prop.

Additionally, the name of ComponentWithClass was misleading as it's used for props rather than components.

This renames ComponentWithClass to PropsWithClassName and starts exporting it.

Note there seems to be a similar problem for `LinkTypes` and possibly others, these aren't fixed here as we also need to be careful to not introduce a breaking change. (There is a small workaround for `LinkTypes` to avoid needing to export it.)