fix(Timeline): Fix duplicated elements in some environments
This fixes problems with duplicated elements in the Timeline component in production builds and/or in older browsers.

These occurred because the Timeline logic was using the name property of functions (the functions being the Timeline components) to reorganise child components, however this failed in a few situations:

- after minification, the function names weren't always unique
- after minification, the functions were changed to anonymous functions, of which names are not supported in older browsers
- very old browsers don't support Function.name at all

(See also https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Function/name.)

This change switches to using a strict equality on the types instead. This compares the functions by reference instead of name, which overcomes those three classes of problems from testing.