feat(ESLintConfigReact): Add exceptions for new rules introduced from deps
The rules added to ignore the default 'no-use-before-define' rule and
instead use @typescript-eslint/no-use-before-define is to address a
known issue with the default Typescript linting rules when used with
React, otherwise is reports an error when you import React.

Arrow body style was also added but there are a great deal many
instances of this pattern used throughout the react components a
test so this is disabled for now but it is advised this pattern is
adopted over time.