fix(DatePicker): Monitor multiple refs
When clicking in the document the `DatePicker`
should close except for when a date is selected
and the input button is clicked. When the input
button is clicked the `DatePicker` will close with
a separate piece of code. Without this fix, the
`DatePicker` is closed and then immediately the
`focus` event fires which reopens it again.