chore(Husky): Upgrade Husky config and fix lint-staged/commitlint
How Husky works changed in version 5 and its scripts were moved from `package.json` to a `.husky` directory.

This meant Husky, lint-staged and commitlint weren't working as we had new versions of these installed but an old Husky configuration.

This upgrades the Husky configuration using https://github.com/typicode/husky-4-to-7 (plus some tweaks) and gets things working again.

Husky hooks now live in the `.husky` directory. Note that when `yarn install` is run in the root directory, a `.husky/_` directory is created. This doesn't get committed to Git as it is automatically .gitignored.