chore(ESLint): Fix linting of cypress directory
It wasn't possible to commit files in the `packages/react-component-library/cypress` directory as lint-staged would try to lint them, and this would fail with the infamous

```
Parsing error: "parserOptions.project" has been set for @typescript-eslint/parser.
The file does not match your project config: cypress/support/index.ts.
The file must be included in at least one of the projects provided
```

 error.

This makes a few changes to fix this following the approaches suggested in:
https://github.com/typescript-eslint/typescript-eslint/blob/master/docs/getting-started/linting/TYPED_LINTING.md#i-get-errors-telling-me-the-file-must-be-included-in-at-least-one-of-the-projects-provided

A separate `tsconfig.eslint.json` TypeScript configuration file that extends `tsconfig.json` is used in order to stop types being generated for Cypress tests and included in the production build.

The `lint` and `lint:ci` commands were also updated to include the `cypress` directory for consistency.