docs(Storybook): Change default story layout
The `fullscreen` layout is awkward for things like form controls, as shadows (and sometimes borders) are out of view.

This changes the default layout from `fullscreen` to the default `padded` layout, and goes through and applies the `fullscreen` layout to components where it makes sense or wasn't getting in the way of borders or shadows.