fix(Storybook): Resolve storybook docs styling in Firefox
Storybook obscures components that have hidden elements (think ContextMenu or DatePicker) when viewed with the Firefox browser.

This made the docs section almost unusable in Firefox. The only solution at this time appears to be overiding some Storybook styles.

https://github.com/storybookjs/storybook/issues/16714
https://github.com/storybookjs/storybook/issues/8928

Included a `ContextMenu` cypress test to help catch the issue with Storybook version bumps.