test(DatePicker): Fix Cypress border assertions in Firefox
CSS assertions on the `border` shorthand property aren't supported in Firefox. Assertions on the constituent components have to be done instead (`border-top-color`, `border-right-color` etc.) which works in Chrome as well.

This adds a custom Chai assertion to do that more easily, and updates the DatePicker and DatePickerE tests to use that.