feat(DatePickerE): Preserve invalid typed dates
This reworks how typed dates are handled so that invalid dates are preserved when the input loses focus.

This also means that the `onChange` callback is now called when an invalid date has been keyed, with an invalid date object (a `Date` with value `NaN`) used to represent an invalid date. (This must be done as otherwise the hosting component will think the field has whatever date was previously there, and this could end up being submitted in the case of a form.)

Similarly, if a disabled date is keyed, `onChange` is still called with this so that the hosting component can display a meaningful error message in line with the design spec.

`startDateValidity` and `endDateValidity` properties have also been added to the `onChange` argument to indicate if a typed date is invalid or disabled.

These changes also fix pasting a date into the field, and manually clearing the field by deleting the date.