fix(Select): Correct label positioning
This fixes a problem where the `Select` label was intermittently not vertically centred. This was happening particularly in Storybook, and was due to react-select's styles sometimes overriding our styles (probably a race condition as it uses emotion for styles).

This fixes that, and tweaks the positioning slightly so that it is better centred.