fix(Fieldset): Improve styling in IE11
This fixes the main styling problems in IE11 when `CheckboxE` and `RadioE` are used in a `Fieldset` or a `FormikGroupE`.

These were happening because IE doesn't support the `unset` keyword (or `initial` or `all`).

It does support `inherit` but that doesn't make sense for many properties.

I've changed replaced these usages of `unset` with what the initial values are according to MDN.