build(IconLibrary): Optimise icon library build time
This optimises the icon library build time by:

- making use of @babel/plugin-transform-runtime to each individual icon component being bloated with Babel helpers
- moving the width, height and fill attribute manipulation to SVGR itself (this avoids them running twice for the ES and CJS builds)
- turning off a Terser option that was really slow and had negligible benefit

The first was the initial trigger of the slow down of the Webpack build. This was made worse after an update to the caniuse database caused rest parameters to be (correctly) transpiled, and a copy of the related babel helpers was being added for each icon.