test(DatePickerE): Fix intermittently failing test
This rewrites and re-enables the skipped DatePickerE test for prop updates that was previously failing in some conditions.

The failures were somehow linked to using Jest fake timers, but the failures continued even if real timers were reactivated just for this test. There was no clear reason for this (the component state was updating correctly, but the expected rerender wasn't happening). The same example also behaves correctly in Storybook.

The test has been amended to use `rerender()` to test updating props in a similar fashion to this example: https://testing-library.com/docs/example-update-props

This is behaving as expected. It can also be verified that the test is updating props and not remounting the component by commenting out this line and observing that the test fails:

https://github.com/defencedigital/mod-uk-design-system/blob/bad6d69b216f80c4c6ef91d20604bf2628d3cea6/packages/react-component-library/src/components/DatePickerE/useSelection.ts#L76