fix(NumberInputE): Allow negative values to be typed or pasted
This updates NumberInputE to allow negative values when `min` is not set or is less than zero.

The input masking behaviour was also improved to eliminate cursor jumping in all browsers.

Note that testing this is difficult in both Cypress and user-event as they both have limitations relating to typing and clipboard operations.