fix(DatePickerE): Allow state to be held externally
This fixes problems with DatePickerE when used with e.g. a form library and the value from the from state is passed to the component.

This is done by splitting the `startDate` and `endDate` props into `startDate`, `initialStartDate`, `endDate` and `endStartDate`. These are analogous to the `value` and `defaultValue` props of `<input>`, where `initialStartDate` and `initialEndDate` only affect the initial values, and `startDate` and `endDate` are always the value of the component if passed.

The component was refactored to use the `useReducer` hook as part of this.