feat(ReactComponentLibrary): Turn on strict null checks
Having strict null checks off means that types implicitly allow null and undefined. This is undesirable as it means:

- types are misleading for users of the component library
- there may be hidden bugs (either due to values being null or undefined unintentionally, or due to null or undefined not being handled)
- there may be missed test cases

This change turns on strict null checks by removing `"strictNullChecks": false` from `tsconfig.json`.

Errors in deprecated components that are about to be removed have been ignored, as there is a separate piece of work in progress to actually remove them.