fix(CheckboxRadioBase): Resolve 'nested-interactive' axe v4 error
This resolves the following error:

https://dequeuniversity.com/rules/axe/4.4/nested-interactive

This was occurring as we effectively had a checkbox within a checkbox, as we had a visually-hidden `<input>` contained within an element with `role="checkbox"`.

The simply removes the roles from the parent elements.

A further improvement would be to make the outer element the `<label>` and turn the child elements that are `<div>`s into `<span>`s. However, this is a larger refactor so I've left it out.