fix(List): Resolve 'aria-required-parent' axe v4 errors
This resolves the following error:

https://dequeuniversity.com/rules/axe/4.4/aria-required-parent

This was due to `listtem` roles not being within a `list` role. (On further investigation, the `group` role being used as a parent could be used to group multiple `listitem`s within a `list`, but isn't a replacement for a `list.)

I've removed the roles completely as `ol` and `li` are already being used.