fix(TabSet): Resolve 'nested-interactive' axe v4 errors
This resolves the following error:

https://dequeuniversity.com/rules/axe/4.4/nested-interactive

This was due to the `tab` containing a `button`. I simply rejigged things so that the `<button>` is now the `tab` (following the same pattern as the GOV.UK Design System).

This also corrects `aria-labelledby` for the tab contents, as this was being set to itself instead of the tab button.