fix(Modal): Resolve 'aria-dialog-name' axe v4 error in stories
This resolves the following error for the Modal stories without a title:

https://dequeuniversity.com/rules/axe/4.4/aria-dialog-name

This was due to the `dialog` not having an accessible name when the Modal had no title set. Perhaps it would be best if titles were required for Modals, but this simply makes it possible to explicitly set `aria-label` for a Modal to avoid any breaking changes.

Note that there was further complexity as Dialog is based on Modal.