fix(IconLibrary): Resolve installation errors
This replaces react-svg-unique-id with https://github.com/laleksiunas/inline-svg-unique-id due to peer dependency problems with the former causing an infinite loop when installing the icon library with the latest npm.

Additionally, `react-svg-unique-id` hasn't been updated in two years, and still has a peer dependency on React 16.

This replaces it with https://github.com/laleksiunas/inline-svg-unique-id which is more up-to-date.