ci(GHA): Skip SonarCloud scan on PRs from forks
Running a SonarCloud scan fails on PRs from forks as it requires secrets. Therefore this skips it for forks (in addition to Dependabot updates).

Running a SonarCloud scan requires secrets to be available and the code being scanned to be checked out, which is not generally recommended for untrusted code. Therefore there is no nice way to make it work.

See also https://portal.productboard.com/sonarsource/1-sonarcloud/c/50-sonarcloud-analyzes-external-pull-request