ci(GHA): Handle Chromatic builds in PRs from forks
This changes how Chromatic is invoked in order to make it compatible with forks.

It is now triggered using the workflow_run event using a pre-built Storybook, so that it can be run in a trusted environment without exposing secrets to untrusted code.

This also fixes the reporting of Chromatic status for normal PRs.

Finally, this change also fixes running of Chromatic for Dependabot PRs as Chromatic was being run against master in this case.