fix(CheckboxRadioBase): Resolve 'nested-interactive' axe v4 error
This resolves the following error:

https://dequeuniversity.com/rules/axe/4.4/nested-interactive

This was occurring as we effectively had a checkbox within a checkbox, as we had a visually-hidden `<input>` contained within an element with `role="checkbox"`.

The removes the roles from the parent elements.

A further problem was that the 0px <input> caused a screen reader used with Firefox to behave oddly. This reorganises things slightly so that it has the same dimensions as`StyledCheckmark`.