feat(GlobalStyleProvider): Handle multiple instances gracefully
This adds logic to GlobalStyleProvider to stop it inserting duplicate styles when it is inserted into a page multiple times.

Note that there are many edge cases in trying to do this, in particular relating to the order different instances are mounted and unmounted. Hence, I'm not convinced trying to do this is a good idea – though I've managed to come up with something that handles those edge cases.