fix(CRATemplate): Fix installation error with React 18.2.0
The CRA template and smoke test starting failing again since the release of React 18.2.0, because the overrides are for react@^18.1.0 while the main React dependency injected by CRA is now react@^18.2.0.

There is a proper way to avoid this using:

```
    "overrides": {
      "react": "$react",
      "react-dom": "$react-dom"
    },
```

However, this is currently not working and results in this error:

```
npm ERR! Unable to resolve reference $react
```

At first I thought this was due to how CRA installs packages, but that doesn't actually seem to be the case; it may be due to https://github.com/npm/cli/issues/4395

For now the only solution seems to be to manually keep it up-to-date, unfortunately (until either `$react` is working or all peer dependency ranges in our tree are good).