fix(Autocomplete): Handle changes to options list after initial render
Changes to the option list after the initial render (e.g. loading options asynchronously) previously didn't work, as a copy of the options was kept in the state. As the options are made up of ReactElements (which also have ReactNodes within their props), it is also difficult to keep track of changes to them.

This changes how things work slightly so the filter state is tracked in the state, rather than the items themselves, which fixes the problem.