fix(Autocomplete): Behave as controlled component
Previously, the Autocomplete `value` prop was only setting the initial value of the component, which was unexpected for a controlled component. In particular, this resulted in unexpected behaviour when the value or children changed externally.

This makes the component controlled when `value` is set so it behaves as expected in these scenarios. `initialValue` was also added for uncontrolled scenarios.