fix(Select): Behave as controlled component
Previously, the Select `value` prop was only setting the initial value of the component, which was unexpected for a controlled component. In particular, this resulted in unexpected behaviour when the value or children changed externally.

This makes the component controlled when `value` is set so it behaves as expected in these scenarios. `initialValue` was also added for uncontrolled scenarios.

This also brings the component in line with recent changes to Autocomplete.