fix(Select): Allow options to be disabled
This exposes missing functionality in Select to allow options to be disabled.

There was no lighter colour in the neutral palette available with sufficient contrast, so the text of non-disabled options was made darker.

A newer versions of Downshift was needed as it contains a bug fix for disabled items. However, newer versions of Downshift are no longer transpiled to ES5. This was worked around by transpiling it in our own bundles.